requirements.txt
numpy
streamlit
openai
python-dotenv